Crate port_io

CPU-level input/output instructions, including inb, outb, etc., and a high level Rust wrapper.

Structs

Port

A readable and writable I/O port over an arbitrary type supporting the InOut interface, which is only u8, u16, or u32.

PortReadOnly

A read-only I/O port over an arbitrary type supporting the InOut interface, which is only u8, u16, and u32.

PortWriteOnly

A write-only I/O port over an arbitrary type supporting the InOut interface, which is only u8, u16, and u32.

Traits

PortIn

This trait is defined for any type which can be read from a port. x86 processors support Port IO for u8, u16 and u32.

PortOut

This trait is defined for any type which can be read from a port. x86 processors support Port IO for u8, u16 and u32.

Functions

inb⚠

Read a u8-sized value from port.

inl⚠

Read a u32-sized value from port.

inw⚠

Read a u16-sized value from port.

outb⚠

Write a u8-sized value to port.

outl⚠

Write a u32-sized value to port.

outw⚠

Write a u8-sized value to port.